Added the backend and frontend parts for the database but still need to test to see if it works
