+++ back/src/VitaWave.WebAPI/Program.cs
@@ -2,7 +2,7 @@
 using VitaWave.Data;
 using VitaWave.WebAPI.Hubs;
 using VitaWave.WebAPI.Notifications;
-
+using VitaWave.DataBase;
 var builder = WebApplication.CreateBuilder(args);
 
 Log.Logger = new LoggerConfiguration()
@@ -37,6 +37,7 @@ builder.Services.AddSignalR();
 
 builder.Services.AddSingleton<DataFacilitator>();
 builder.Services.AddSingleton<NotificationHandler>();
+builder.Services.AddSingleton<DataBase>();
 
 
 var app = builder.Build();
