+++ front/package.json
@@ -12,5 +12,8 @@
   "dependencies": {
     "@microsoft/signalr": "^8.0.0",
     "electron": "^28.0.0"
+  },
+  "devDependencies": {
+    "eslint": "^9.39.1"
   }
 }
